Manually close the payment if the configuration requested it
Since API v1.9 the underlying library used for requesting CSOB API
doesn't read the closePayment config globally and reads it only
from the $newPayment instance.

Unfortunately there's no way how to tell $newPayment that it should
close and we end up with unclaimed payments.

If the caller configured the omnipay library to close the payment,
this will enforce the closure.